Add copy constructor to Creature to fix hostility issues.
Prior to this, when creatures were being copied, their shared pointer to a
decision strategy was also being copied.  This was problematic, because the
shared_ptr meant that hostility information was being shared between creatures.
The implementation of the copy constructor/operator=/swap functions ensures
that a proper new copy with a distinct DecisionStrategyPtr is made.  In
combination with making a copy off the creature template (which will never
have any hostility information), peaceful creatures are now possible.