Add CommandFactory serialization, and update Display serialization.
The Display serialization was just checking ClassIdentifier values with
if-statements, and instantiating based on that.  I've updated the approach
so that maps are used instead.  The maps contain shared_ptrs, and the
objects wrapped by the shared_ptr always contain a clone() method.  So,
when a value is needed, we can essentially do shared_ptr(map[x]->clone).