Fix savefile name hashing bug.
The savefile name is generated by combining the user name with the creature
name, and creating a boost hash based off of that.  Prior to my changes,
the creature name was just concatenated with the user name, producing a
name to be hashed like "JulianThrundarr".  The problem is that if there
were a user called "Julia" and a character named "nThrundarr", it would
generate the same string to hash.  So, I added an additional underscore
after the user name.