Tile features weren't being serialized properly - fixed.  Features also weren't being set on tiles after deserialization.
Non-custom maps (those generated with a GeneratorPtr) didn't have IDs set properly, so the player could sometimes end up on the wrong map.  This would cause an infinite loop when trying to enter a map, but only sometimes...

Added a check to force an NPC's action to advance the turn.  There's an infinite loop hiding somewhere in the monster movement/pathfinding code, so this is just a band-aid while I look for the actual cause of the issue.